Resolved the following issues - Fixes #39 - Fixes #40 added methods to check if time t has elapsed since winning last bid

+++ src/main/java/utils/simulation/Timer.java
@@ -66,7 +66,7 @@ public class Timer {
 
         if (second >= 60.0) // Method for outputting seconds
         {
-            minute += (int) (second / 60);
+            minute += second / 60;
             second %= 60;
         }
